backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
boto3==1.26.142
botocore==1.29.142
gunicorn==20.1.0
requests==2.30.0
pytest==7.3.1
Werkzeug==2.3.4